+++ src/components/features/mypage/MyReserveCalendar/MyReserveCalendar.tsx
@@ -1,5 +1,5 @@
 import MyCalendar from './MyCalendar/MyCalendar';
-import ScheduleBox from './ScheduleBox';
+import ScheduleBox from './ScheduleBox/ScheduleBox';
 
 const MyReserveCalendar = () => {
   return (
